requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
Django==5.1.6
django-db-geventpool==4.0.7
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
kombu==5.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
sentry-sdk==2.22.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
psycopg2-binary
python-decouple
